使用 autofix 修复多个问题 (#80)
* ci: update .deepsource.toml

* refactor: refactor for loop into `any` call

It is recommended to use `any` for this loop, as it will make your code smaller and more readable.

* refactor: remove unused imports

An object has been imported but is not used anywhere in the file.
It should either be used or the import should be removed.

* style: format code with Black

This commit fixes the style issues introduced in fb05a11 according to the output
from Black.

Details: None

* refactor: refactor unnecessary `else` / `elif` when `if` block has a `return` statement

The use of `else` or `elif` becomes redundant and can be dropped if the last statement under the leading `if` / `elif` block is a `return` statement.
In the case of an `elif` after `return`, it can be written as a separate `if` block.
For `else` blocks after `return`, the statements can be shifted out of `else`. Please refer to the examples below for reference.

Refactoring the code this way can improve code-readability and make it easier to maintain.

* refactor: remove unnecessary f-string

It is wasteful to use `f-string` mechanism if there are no expressions to be extrapolated. It is recommended to use regular strings instead.

* refactor: remove redundant `None` default

It is unnecessary to provide `None` as the default value when the key is not present in the dictionary as `get` implicitly returns `None`.  Not preferred:

* refactor: iterate dictionary directly

Consider iterating the dictionary directly instead of calling `.keys()`.
Using `for key in dictionary` would always iterate the dictionary keys.

* refactor: refactor unnecessary `else` / `elif` when `if` block has a `raise` statement

`raise` causes control flow to be disrupted, as it will exit the block.
It is recommended to check other conditions using another `if` statement, and get rid of `else` statements as they are unnecessary.

* refactor: use identity check for comparison to a singleton

Comparisons to the singleton objects, like `True`, `False`, and `None`, should be done with identity, not equality. Use `is` or `is not`.

* refactor: refactor unnecessary `else` / `elif` when `if` block has a `continue` statement

The use of `else` or `elif` becomes redundant and can be dropped if the last statement under the leading `if` / `elif` block is a `continue` statement.
In the case of an `elif` after `continue`, it can be written as a separate `if` block.
For `else` blocks after `continue`, the statements can be shifted out of `else`. Please refer to the examples below for reference.

Refactoring the code this way can improve code-readability and make it easier to maintain.

* refactor: remove empty block of code

In most cases, an empty body of `for`, `while` or `if` implies some piece of code is missing.
Such empty block must be either filled or removed.

* refactor: replace multiple `==` checks with `in`

To check if a variable is equal to one of many values, combine the values into a tuple and check if the variable is contained `in` it instead of checking for equality against each of the values.
This is faster, less verbose, and more readable.

* refactor: remove unnecessary statement

The expression or statement is pointless and will not affect the program's execution. It should be removed.

* refactor: remove unnecessary except handlers

Using `raise` as the first or only operator of an except handler is useless because it raises back the exception immediately.
Remove the `raise` operator or the entire `try-except-raise` block.

* refactor: remove unnecessary use of comprehension

It is unnecessary to use a comprehension just to loop over the `iterable` and create a `list`/`set`/`dict` out of it. Python has a specialized set of tools for this task: the `list`/`set`/`dict` constructors, which are faster and more readable.

* refactor: refactor unnecessary `else` / `elif` when `if` block has a `return` statement

The use of `else` or `elif` becomes redundant and can be dropped if the last statement under the leading `if` / `elif` block is a `return` statement.
In the case of an `elif` after `return`, it can be written as a separate `if` block.
For `else` blocks after `return`, the statements can be shifted out of `else`. Please refer to the examples below for reference.

Refactoring the code this way can improve code-readability and make it easier to maintain.

* refactor: refactor useless `else` block in the loop

When a loop specifies no break statement, the else clause will always execute when the loop sequence is empty, thus making it useless.
It is recommended to have the statements under else in the same scope as the loop itself.
If the else clause does not always execute at the end of a loop clause, then the code should add a break statement within the loop block.

* style: format code with Black

This commit fixes the style issues introduced in f43a3a4 according to the output
from Black.

Details: https://github.com/wling-art/ToolDelta/pull/18

* refactor: remove reimported module

A module or an import name is reimported multiple times. This can be confusing and should be fixed.
Please refer to the occurrence message to see the reimported name and the line number where it was imported for the first time.

* refactor: replace `StopIteration` with `return`

According to [PEP479](https://www.python.org/dev/peps/pep-0479/), the raise of `StopIteration` to end the loop of a generator may lead to hard to find bugs. This PEP specifies that `raise StopIteration` has to be replaced by a simple `return` statement.

* refactor: replace range(len(...)) with enumerate(...)

Using `range(len(...))` is not pythonic. Python does not have not index-based loops. Instead, it uses collection iterators.  Python has a built-in method `enumerate` which adds a counter to an iterable.

---------

Co-authored-by: deepsource-io[bot] <42547082+deepsource-io[bot]@users.noreply.github.com>
Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>