使用 autofix 自动修复 更多的问题 (#82)
* refactor: group imports from same package

Imports for the same package are scattered and not grouped together. It is recommended to keep the imports from the same package together. It makes the code easier to read.

* refactor: put docstring into a single line

If a docstring fits in a single line (72 characters according to PEP8), it is
recommended to have the quotes on the same line.

* refactor: remove empty block of code

In most cases, an empty body of `for`, `while` or `if` implies some piece of code is missing.
Such empty block must be either filled or removed.

* refactor: remove unused nested definitions

This issue concerns that one or more nested classes/functions have been declared but are nowhere used in the function.

* refactor: merge collapsible `if` statements

Nested `if` statements can be collapsed into a single `if` statement by separating their condition using `and` operator.
Merging collapsible `if` statements increases the code's readability.

* refactor: remove blank lines after docstring

There shouldn't be any blank lines after the function docstring.
Remove the blank lines to fix this issue.

* refactor: remove unnecessary `del` statement from local scope

Passing a local variable to a `del` statement results in that variable being removed from the local namespace. When exiting a function all local variables are deleted, so it is unnecessary to explicitly delete variables in such cases.

* refactor: `raise NotImplementedError` in empty functions

The function/method has been left empty here, without any comment or docstring. This can cause confusion later on why this was left empty.

---------

Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>