提升代码质量以及增加易读性 (#146)
* refactor: iterate dictionary directly

Consider iterating the dictionary directly instead of calling `.keys()`.
Using `for key in dictionary` would always iterate the dictionary keys.

* refactor: remove unnecessary whitespace

Blank lines should not contain any tabs or spaces.

* refactor: replace multiple `==` checks with `in`

To check if a variable is equal to one of many values, combine the values into a tuple and check if the variable is contained `in` it instead of checking for equality against each of the values.
This is faster, less verbose, and more readable.

* refactor: group imports from same package

Imports for the same package are scattered and not grouped together. It is recommended to keep the imports from the same package together. It makes the code easier to read.

* refactor: remove assert statement from non-test files

Usage of `assert` statement in application logic is discouraged. `assert` is removed with compiling to optimized byte code. Consider raising an exception instead. Ideally, `assert` statement should be used only in tests.

* refactor: put docstring into a single line

If a docstring fits in a single line (72 characters according to PEP8), it is
recommended to have the quotes on the same line.

* refactor: remove commented out code

It is recommended to remove any commented code in your codebase.

* refactor: add newline at end of file

It is recommended to put a newline at the end of the file.

* refactor: refactor unnecessary `else` / `elif` when `if` block has a `return` statement

The use of `else` or `elif` becomes redundant and can be dropped if the last statement under the leading `if` / `elif` block is a `return` statement.
In the case of an `elif` after `return`, it can be written as a separate `if` block.
For `else` blocks after `return`, the statements can be shifted out of `else`. Please refer to the examples below for reference.

Refactoring the code this way can improve code-readability and make it easier to maintain.

* refactor: refactor unnecessary `else` / `elif` when `if` block has a `return` statement

The use of `else` or `elif` becomes redundant and can be dropped if the last statement under the leading `if` / `elif` block is a `return` statement.
In the case of an `elif` after `return`, it can be written as a separate `if` block.
For `else` blocks after `return`, the statements can be shifted out of `else`. Please refer to the examples below for reference.

Refactoring the code this way can improve code-readability and make it easier to maintain.

* refactor: refactor unnecessary `else` / `elif` when `if` block has a `raise` statement

`raise` causes control flow to be disrupted, as it will exit the block.
It is recommended to check other conditions using another `if` statement, and get rid of `else` statements as they are unnecessary.

* refactor: refactor unnecessary `else` / `elif` when `if` block has a `continue` statement

The use of `else` or `elif` becomes redundant and can be dropped if the last statement under the leading `if` / `elif` block is a `continue` statement.
In the case of an `elif` after `continue`, it can be written as a separate `if` block.
For `else` blocks after `continue`, the statements can be shifted out of `else`. Please refer to the examples below for reference.

Refactoring the code this way can improve code-readability and make it easier to maintain.

* refactor: use identity check for comparison to a singleton

Comparisons to the singleton objects, like `True`, `False`, and `None`, should be done with identity, not equality. Use `is` or `is not`.

* refactor: refactor unnecessary `else` / `elif` when `if` block has a `return` statement

The use of `else` or `elif` becomes redundant and can be dropped if the last statement under the leading `if` / `elif` block is a `return` statement.
In the case of an `elif` after `return`, it can be written as a separate `if` block.
For `else` blocks after `return`, the statements can be shifted out of `else`. Please refer to the examples below for reference.

Refactoring the code this way can improve code-readability and make it easier to maintain.

* refactor: remove unnecessary lambda expression

A lambda that calls a function without modifying any of its parameters is unnecessary. Python functions are first-class objects and can be passed around in the same way as the resulting lambda. It is recommended to remove the lambda and use the function directly.

* refactor: `raise NotImplementedError` in empty functions

The function/method has been left empty here, without any comment or docstring. This can cause confusion later on why this was left empty.

* refactor: change methods not using its bound instance to staticmethods

The method doesn't use its bound instance. Decorate this method with `@staticmethod` decorator, so that Python does not have to instantiate a bound method for every instance of this class thereby saving memory and computation. Read more about staticmethods [here](https://docs.python.org/3/library/functions.html#staticmethod).

* refactor: remove implicit `object` from the base class

The class is inheriting from `object`, which is implicit under Python 3 , hence can be safely removed from bases.  Not preferred:

* refactor: remove unnecessary f-string

It is wasteful to use `f-string` mechanism if there are no expressions to be extrapolated. It is recommended to use regular strings instead.

* refactor: import packages in separate lines

Imports from different modules should be on different lines.

* refactor: merge collapsible `with` statements

Nested `with` statements can be merged into a single compound `with` statement.

This issue is raised only when there are no other statements between the nested `with` statements.
Merging collapsible `with` statements can decrease cognitive complexity, improving code readability.

* refactor: remove redundant `None` default

It is unnecessary to provide `None` as the default value when the key is not present in the dictionary as `get` implicitly returns `None`.  Not preferred:

---------

Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>